feature : calculate cleaning delay
