+++ src/test/java/align/AlignCleaningServiceTest.java
@@ -1,0 +1,5 @@
+package align;
+
+public class AlignCleaningServiceTest {
+
+}
