+++ src/test/java/service/builder/ScheduleBuilderTest.java
@@ -5,7 +5,7 @@ import org.acme.foodpackaging.domain.PackagingSchedule;
 import org.acme.foodpackaging.domain.Product;
 import org.acme.foodpackaging.record.InitData;
 import org.acme.foodpackaging.repository.jobs.JobRepository;
-import org.acme.foodpackaging.service.builder.AlignSolutionService;
+import org.acme.foodpackaging.service.align.AlignSolutionService;
 import org.acme.foodpackaging.service.jobs.JobRefreshService;
 import org.acme.foodpackaging.service.jobs.JobService;
 import org.acme.foodpackaging.service.products.ProductService;
@@ -32,7 +32,6 @@ class ScheduleBuilderTest {
 
     @InjectMocks
     ScheduleBuilder builder;
-
     @Mock
     JobService jobService;
     @Mock
@@ -75,8 +74,7 @@ class ScheduleBuilderTest {
         verify(lineService).initLineStartEnd(schedule);
         verify(productService).getProductList(schedule);
 
-        verify(alignSolutionService).alignByFactDuration(schedule);
-        verify(alignSolutionService).alignLineStartByFact(schedule);
+        verify(alignSolutionService).align(schedule);
     }
 
     @Test
